+++ location.h
@@ -8,7 +8,7 @@ public:
     double latitude, longitude;
     Location(double lat, double lon);
     double distanceTo(const Location &other) const;
-    double travelTimeTo(const Location &other) const;
+    int travelTimeTo(const Location &other) const;
 };
 
 #endif
